requirements.txt
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
filelock==3.20.3
idna==3.11
pycparser==2.23
requests==2.32.5
urllib3==2.6.3
python-dotenv==1.2.1
pip==25.3
